1. 修改SDPackage结构 2. 修改ReadCallback的参数类型为SDPackage* 3. SDMaster增加处理远程服务发现的逻辑 4. 本地sd_service注册服务时加入自身IP和Port信息 5. 修复所有的编译告警

+++ include/common/transfer.h
@@ -2,6 +2,7 @@
 #define TRANSFER_H_
 
 #include <sys/socket.h>
+#include <sys/types.h>
 
 /// @brief 循环接收data_size字节的数据，还可以用flag为MSG_WAITALL的普通recv代替
 /// @param fd socket描述符
@@ -31,6 +32,6 @@ int sendn(int fd, void* vptr, int data_size, int flag);
 /// @param msg 存储接收到数据的结构体指针
 /// @param flag 无论传入什么都指定为MSG_WAITALL
 /// @return 返回实际接收数据的字节数
-int recvmsgn(int sockfd, msghdr* msg, int flag);
+ssize_t recvmsgn(int sockfd, msghdr* msg, int flag);
 
 #endif //TRANSFER_H_
